pyproject.toml
[tool.poetry]
name = "shioajicaller"
version = "0.1.4.0"
description = "shioaj warp caller"
authors = ["Steve Lo <info@sd.idv.tw>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
shioaji = "^0.3.3.dev2"
python-dotenv = "^0.19.0"
redis = "^3.5.3"
websockets = "^10.0"
uvloop = "^0.16.0"
aioredis = "2.0.0"

[tool.poetry.dev-dependencies]
python = "^3.7"
shioaji = "^0.3.3.dev2"
python-dotenv = "^0.19.0"
redis = "^3.5.3"
websockets = "^10.0"
uvloop = "^0.16.0"
aioredis = "2.0.0"

[tool.poetry.scripts]
shioajicaller = "shioajicaller.cli:run"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
